Fix review comments and small refactoring
